backend/pyproject.toml
[project]
name = "AICORO"
version = "0.1.0"
dependencies = [
    "fastapi>=0.104.1",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "uvicorn[standard]>=0.24.0",
    "motor>=3.3.0",
    "pydantic>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src/AICORO"]

